+++ fiddles/angular2-cli/fiddle-0037-EntAgGridStateManagement/src/app/components/toolbar/toolbar.component.ts
@@ -1,0 +1,12 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+    selector: 'app-toolbar',
+    templateUrl: './toolbar.component.html',
+    styleUrls: ['./toolbar.component.css']
+})
+export class ToolbarComponent {
+
+    @Input() componentId: string;
+
+}