+++ CHANGELOG.md
@@ -2308,3 +2308,11 @@ Alternately, the _Daily Diary of my Programming Dreams_.
 * **angular Cli**
 
     1.  Added [fiddles/angular2-cli/fiddle-0036-EntAgGridToolTips](fiddles/angular2-cli/fiddle-0036-EntAgGridToolTips)
+
+
+### 347_angular_37 ~ [Issue 347](https://github.com/bradyhouse/house/issues/347)
+
+* **angular Cli**
+
+    1.  Added [fiddles/angular2-cli/fiddle-0037-EntAgGridStateManagement](fiddles/angular2-cli/fiddle-0037-EntAgGridStateManagement)
+